start inremeanting device id from 1
starting at 0 is not good (it indicates that it failed or something), so
start from 1

also speed in full u64, just in case

+++ src/handler/device_list.rs
@@ -11,7 +11,7 @@ use tokio::io::AsyncWriteExt;
 #[must_use]
 pub fn create_device_connected_plist(
     id: u64,
-    speed: u32,
+    speed: u64,
     device_address: u8,
     product_id: u16,
     serial_number: String,
